requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
drf-yasg==1.21.8
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyjwt==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
gunicorn==21.2.0
django-cors-headers==4.3.1
loguru==0.7.2
django-filter==23.5
django-debug-toolbar==4.3.0
Pillow==10.2.0
whitenoise==6.6.0
